Implemented simple loading
I'm not quite satisfied with the way the loading is handled right now. The Option Frontman is responsible for processing the indexes in the save file, which are pretty loosely interpretable by themselves. But that shouldn't be it's task in the first place. There should probably be a class responsible for interpreting and confirming the initial state of the character creator.

On top of that. While the blackboard is a good idea, it needs to be implemented better. Preferably as a Singleton. I have a useful class for that tho.

Finally. What if there will be different types of options? Like a slider for example. That can't have an index. How would the program know on which Active Body Part class to apply that too, or which Option Handler? Further research required. :)